src/requirements.txt
mido
pygame
fluidsynth
pyfluidsynth
paho-mqtt